* refactor (step3) : class 명 변경
- 과제 설명에서 표현된 명칭으로 변경
- CarRacing -> RacingGame
- RacingGame -> RacingMain
